requirements.txt
pandas
numpy
scikit-learn
xgboost
tensorflow
matplotlib
seaborn
fuzzywuzzy
python-Levenshtein
nltk
spacy
google-generativeai
python-dotenv
jupyter
ipykernel